Rendering Start
Started on backend_rendering and some front_end rendering. Current code contains some temporary code to see if the z buffer is working with some 3D rendering. This commit is a bit of a frankestien because I forgot to regularly do this. The commit all contains some style changes to some of the functions and structs so that the code is compliant with the google code style.